Use pytest-xdist and -n auto option to run tests
This solves the problem when test_spark_config_precedence_in_transformation_2
was not working within a single session. The issue is that spark session from previous tests
is passing to the next tests. Even thought the spark session fixture has a function scope.
Thus the new config based on yaml was not set. Running the test separately was solving the
problem.

Previously the following items were tried
without the success:
- Modifying the spark fixture by adding the teardown instructions. After each
test spark_session.stop() and spark_session.currentContext.stop() was called
- setting SparkSession._currentContext to None at the teardown

The solution with pytest-xdist plugin and -n auto means that each test
runs in a separate process separately, what gave two benefits:
- Resolves the problem with the spark session migration
- speed up the test run significantly